refactor: move account verification to a helper, session-scope dedup state
