+++ tests/benchmark/helper/account_creator.py
@@ -1,7 +1,7 @@
 """Benchmark target accounts of various kinds for creation and location.."""
 
 from abc import ABC, abstractmethod
-from collections.abc import Callable, Hashable, Iterable, Iterator
+from collections.abc import Hashable, Iterable, Iterator
 from dataclasses import dataclass
 from enum import Enum, auto
 from typing import ClassVar, Self
@@ -20,6 +20,11 @@ from execution_testing import (
 )
 from execution_testing.forks import Osaka
 
+from tests.benchmark.helper.account_verification import (
+    AccountExpectation,
+    register_target_range,
+)
+
 DEFAULT_CODE_SIZE = Osaka.max_code_size()
 
 ADDRESS_MASK = (1 << 160) - 1
@@ -29,71 +34,6 @@ ADDRESS_MASK = (1 << 160) - 1
 EXISTING_EOA_BASE = 0x1000
 # An address range that is never funded.
 NON_EXISTING_BASE = keccak256(b"random")
-
-
-@dataclass(frozen=True)
-class AccountExpectation:
-    """
-    Expected on-chain shape of a snapshot-predeployed target.
-
-    Verified at `start_block`. Defaults skipped.
-    `is_contract`: nonce >= 1. CREATE2: address binds code.
-    `code_prefix`: on-chain code must start with given bytes.
-    """
-
-    is_existing_account: bool = True
-    is_contract: bool = False
-    min_balance: int | None = None
-    code_prefix: bytes | None = None
-
-    def register(
-        self, pre: Alloc, address: Address, *, label: str | None = None
-    ) -> None:
-        """
-        Register verification expectation for address on pre-allocation.
-        """
-        pre.expect_account_state(
-            address,
-            is_existing_account=self.is_existing_account,
-            is_contract=self.is_contract,
-            min_balance=self.min_balance,
-            code_prefix=self.code_prefix,
-            label=label,
-        )
-
-
-_verified_target_counts: dict[Hashable, int] = {}
-
-
-def register_target_range(
-    pre: Alloc,
-    *,
-    key: Hashable,
-    count: int,
-    expectation: AccountExpectation,
-    addresses: Callable[[int, int], Iterable[Address]],
-    label: str | None = None,
-    full: bool = False,
-) -> None:
-    """
-    Register targets [0, count) for verification, deduped across session.
-
-    With *full*, register every target; otherwise only first and last.
-    No-op for allocations without a snapshot.
-    """
-    start = _verified_target_counts.get(key, 0)
-    if count <= start:
-        return
-    if full or count - start <= 2:
-        targets: Iterable[Address] = addresses(start, count)
-    else:
-        targets = [
-            *addresses(start, start + 1),
-            *addresses(count - 1, count),
-        ]
-    for address in targets:
-        expectation.register(pre, address, label=label)
-    _verified_target_counts[key] = count
 
 
 class AccountMode(Enum):
@@ -471,6 +411,7 @@ class AccountCreator:
         pre: Alloc,
         count: int,
         *,
+        verified_accounts: dict[Hashable, int],
         full: bool = False,
         label: str | None = None,
     ) -> None:
@@ -483,6 +424,7 @@ class AccountCreator:
             addresses=lambda start, stop: self.target_addresses(
                 range(start, stop)
             ),
+            verified_accounts=verified_accounts,
             label=label or self.mode.name,
             full=full,
         )
